fix: Challenge 결과/히스토리/현황 응답의 금액 필드를 Long으로 전환

+++ src/main/java/Hampouch/server/domain/challenge/dto/CurrentChallengeResponse.java
@@ -59,7 +59,7 @@ public record CurrentChallengeResponse(
             int successDays,
             int overDays,
             int currentStreak,
-            int savedAmountSoFar
+            long savedAmountSoFar
     ) {
     }
